correction de commission et ajoute d'un methode pour supprimer account

+++ src/main/java/com/paymybud/backend/dto/TransfertDTO.java
@@ -1,4 +1,4 @@
 package com.paymybud.backend.dto;
 
-public record TransfertDTO(Integer amount, String message, String receiverEmail) {
+public record TransfertDTO(Double amount, String message, String receiverEmail) {
 }
